+++ openraft/Cargo.toml
@@ -27,6 +27,7 @@ serde           = { workspace = true, optional = true }
 serde_json      = { workspace = true, optional = true }
 tempfile        = { workspace = true, optional = true }
 monoio          = { workspace = true, optional = true }
+local-sync      = { workspace = true, optional = true }
 thiserror       = { workspace = true }
 tokio           = { workspace = true }
 tracing         = { workspace = true }
@@ -79,8 +80,9 @@ compat = []
 compat-07 = ["compat", "serde", "dep:or07", "compat-07-testing"]
 compat-07-testing = ["dep:tempfile", "anyhow", "dep:serde_json"]
 
-monoio = ["dep:monoio", "singlethreaded"]
-monoio-sync = ["dep:monoio", "monoio/sync"]
+# Turn on monoio Runtime with singlethreaded monoio tasks
+monoio = ["dep:monoio", "singlethreaded", "dep:local-sync"]
+# monoio-sync = ["dep:monoio", "monoio/sync"]
 
 # Allows an application to implement a custom the v2 storage API.
 # See `openraft::storage::v2` for more details.
