+++ openraft/src/async_runtime/tokio.rs
@@ -2,13 +2,10 @@ use std::fmt::Debug;
 use std::future::Future;
 use std::time::Duration;
 
+use crate::async_runtime::AsyncOneshotSendExt;
 use crate::AsyncRuntime;
 use crate::Instant;
 use crate::OptionalSend;
-
-/// `Tokio` is the default asynchronous executor.
-#[derive(Debug, Default)]
-pub struct TokioRuntime;
 
 pub type TokioInstant = tokio::time::Instant;
 
@@ -19,6 +16,12 @@ impl Instant for tokio::time::Instant {
     }
 }
 
+/// `Tokio` is the default asynchronous executor.
+#[derive(Debug, Default, PartialEq, Eq)]
+pub struct TokioRuntime;
+
+pub struct TokioOneShotSender<T: OptionalSend>(pub tokio::sync::oneshot::Sender<T>);
+
 impl AsyncRuntime for TokioRuntime {
     type JoinError = tokio::task::JoinError;
     type JoinHandle<T: OptionalSend + 'static> = tokio::task::JoinHandle<T>;
@@ -27,6 +30,9 @@ impl AsyncRuntime for TokioRuntime {
     type TimeoutError = tokio::time::error::Elapsed;
     type Timeout<R, T: Future<Output = R> + OptionalSend> = tokio::time::Timeout<T>;
     type ThreadLocalRng = rand::rngs::ThreadRng;
+    type OneshotSender<T: OptionalSend> = TokioOneShotSender<T>;
+    type OneshotReceiver<T: OptionalSend> = tokio::sync::oneshot::Receiver<T>;
+    type OneshotReceiverError = tokio::sync::oneshot::error::RecvError;
 
     #[inline]
     fn spawn<T>(future: T) -> Self::JoinHandle<T::Output>
@@ -72,5 +78,25 @@ impl AsyncRuntime for TokioRuntime {
     #[inline]
     fn thread_rng() -> Self::ThreadLocalRng {
         rand::thread_rng()
+    }
+
+    #[inline]
+    fn oneshot<T>() -> (Self::OneshotSender<T>, Self::OneshotReceiver<T>)
+    where T: OptionalSend {
+        let (tx, rx) = tokio::sync::oneshot::channel();
+        (TokioOneShotSender(tx), rx)
+    }
+}
+
+impl<T: OptionalSend> AsyncOneshotSendExt<T> for TokioOneShotSender<T> {
+    #[inline]
+    fn send(self, t: T) -> Result<(), T> {
+        self.0.send(t)
+    }
+}
+
+impl<T: OptionalSend> Debug for TokioOneShotSender<T> {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.debug_tuple("TokioSendWrapper").finish()
     }
 }
